+++ common/binary_search_tree/base/find_by_order.h
@@ -4,13 +4,13 @@
 
 namespace bst {
 template <class TNode>
-inline TNode* FindByOrder(TNode* root, unsigned order_index) {
+inline TNode* FindByOrder(TNode* root, size_t order_index) {
   static_assert(TNode::TInfo::has_size, "info should contain size");
   if (!root) return nullptr;
   if (order_index >= root->info.size) return nullptr;
   for (TNode* node = root; node;) {
     node->ApplyAction();
-    unsigned ls = (node->l ? node->l->info.size : 0);
+    size_t ls = (node->l ? node->l->info.size : 0);
     if (order_index < ls) {
       node = node->l;
     } else if (order_index == ls) {
